+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ##### git
 - Add compatibility with browserpass via `--pass-compat=browserpass`
 - Add compatibility mode `username` for a `username:` prefix
+- Add `--pass-cmd` to allow specifying pass's location or script name.
 
 ##### 0.7.0
 - Fix PK11 slot memory leak