+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add compatibility mode `username` for a `username:` prefix
 - Add `--pass-cmd` to allow specifying pass's location or script name.
 - Using `--pass-prefix=''` prevents creation of a prefix: `web/address/...` becomes `address/...`
+- Fix an encoding bug due to non-ASCII characters leading to a user's profile path
 
 ##### 0.7.0
 - Fix PK11 slot memory leak