+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 #### Changelog
 
-##### git
+##### 0.8.0+git
 - Add compatibility with browserpass via `--pass-compat=browserpass`
 - Add compatibility mode `username` for a `username:` prefix
 - Add `--pass-cmd` to allow specifying pass's location or script name.